feat: create campaign initial implementation done
